+++ test/integration/TransactionTypes/AuthorizationTest.php
@@ -21,9 +21,9 @@
  *
  * @author  Simon Gabriel <development@heidelpay.com>
  *
- * @package  heidelpayPHP/test/integration
+ * @package  heidelpayPHP/test/integration/transaction_types
  */
-namespace heidelpayPHP\test\integration;
+namespace heidelpayPHP\test\integration\TransactionTypes;
 
 use heidelpayPHP\Exceptions\HeidelpayApiException;
 use heidelpayPHP\Resources\AbstractHeidelpayResource;